move notifications_test to integration_test
