--- packages/patrol/example/integration_test/permissions/notifications_test.dart
@@ -1,28 +1,0 @@
-import '../common.dart';
-
-void main() {
-  patrol(
-    'taps on notification',
-    ($) async {
-      await createApp($);
-      await $('Open notifications screen').tap();
-
-      if (await $.native.isPermissionDialogVisible()) {
-        await $.native.grantPermissionWhenInUse();
-      }
-
-      await $('Show in a few seconds').tap();
-      await $.native.pressHome();
-      await $.native.openNotifications();
-
-      // wait for notification to show up
-      await Future<void>.delayed(const Duration(seconds: 5));
-
-      await $.native.tapOnNotificationBySelector(
-        Selector(textContains: 'Someone liked'),
-      );
-
-      await $('Tapped notification with ID: 1').waitUntilVisible();
-    },
-  );
-}
